requirements.txt
google-cloud-speech==2.21.0
google-cloud-translate==2.0.1